Added scripts for switching Java versions and for deployment.
The 'change_java.py' script allows the user to switch between different installed Java versions, updating the JAVA_HOME in the user's .bashrc file. The 'deploy.sh' script performs the deployment of a .war file to the WebLogic server, looking for the .war file in the target directory and undeploying the existing application before deploying the new .war.